+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Although Moulti's Python packages, modules and functions are obviously available
 
 - Fix "next line color" heuristic for step widgets.
 - Fix detection of markup errors.
+- New layout for button questions: there should no longer be any hidden/unreachable button.
 
 ### Removed